make SYSCALL_DEFINE n generated wrappers to asmlinkage

+++ arch/x86/include/asm/syscalls.h
@@ -39,8 +39,8 @@ long sys_sigaltstack(const stack_t __user *, stack_t __user *,
 
 
 /* kernel/tls.c */
-asmlinkage int sys_set_thread_area(struct user_desc __user *);
-asmlinkage int sys_get_thread_area(struct user_desc __user *);
+asmlinkage long sys_set_thread_area(struct user_desc __user *);
+asmlinkage long sys_get_thread_area(struct user_desc __user *);
 
 /* X86_32 only */
 #ifdef CONFIG_X86_32
